requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==41.0.7
dnspython==2.8.0
email-validator==2.1.0
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Mail==0.9.1
fpdf==1.7.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.2
lxml==6.0.2
MarkupSafe==3.0.3
mysql-connector-python==8.1.0
numpy==2.4.2
opencv-python==4.8.1.78
openpyxl==3.1.2
Pillow==10.1.0
protobuf==4.21.12
pycparser==3.0
pypng==0.20220715.0
pyserial==3.5
python-docx==1.1.0
python-dotenv==1.0.0
pyzbar==0.1.9
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==2.3.7
waitress==3.0.2
Flask-Limiter==3.5.0
pyotp==2.9.0
qrcode==7.4.2
sentry-sdk[flask]==1.40.6
